Notify users when component doesn't exists
Before this change, when users specified non existing component,
the application would finish without any errors.

This change makes sure users know they specified non-existing
component and also prints the available components from which
users can choose.

This change also creates classes for Component and Distro for two
reasons:
  1. Sharing code across different znoyder sub-commands
  2. They will be used in the future to modelize components
     and distros
